+++ BookBridge/BookBridge/ViewModel/MyPageViewModel.swift
@@ -42,7 +42,7 @@ extension MyPageViewModel {
 
 extension MyPageViewModel {
     func getDownLoadImage() {
-        print("유져: \(userManager.user?.profileURL)")
+        print("유저: \(userManager.user?.profileURL)")
         print("userSaveImage: \(userSaveImage)")
         if userSaveImage.0 != userManager.user?.profileURL ?? ""{
             guard let urlString = userManager.user?.profileURL else { return }
