+++ BookBridge/BookBridge/Views/MyPage/SettingView.swift
@@ -8,6 +8,8 @@
 import SwiftUI
 
 struct SettingView: View {
+    @StateObject var inquiryVM = InquiryViewModel()
+    
     @Environment(\.dismiss) private var dismiss
     @State private var showingLogoutAlert = false
     @State private var showingLoginView = false
@@ -58,7 +60,7 @@ struct SettingView: View {
             )
             
             NavigationLink {
-                InquiryView()
+                InquiryView(inquiryVM: inquiryVM)
             } label: {
                 HStack {
                     Text("문의 및 건의사항")
